refactor: HTTP 네트워크 통신 디렉토리 트리 개선 #30
- API 모델별 별도 파일 생성
